Improve code quality for module list commands
